% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cities_cases}
\alias{cities_cases}
\title{Covid19 cases in Saudi cities}
\usage{
cities_cases()
}
\value{
data frame with cities covid19 cases
}
\description{
use this to view covid19 cases in cities in Saudi Arabia
}
