% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cities_covid19_first}
\alias{cities_covid19_first}
\title{first cases in cities}
\usage{
cities_covid19_first()
}
\value{
df with first reported cases in each city
}
\description{
This function returns cities with their first reported cases
}
