% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cities_covid19_growth}
\alias{cities_covid19_growth}
\title{Growth of covid 19 cases in Saudi cities}
\usage{
cities_covid19_growth(freqeuncy = "weekly")
}
\arguments{
\item{freqeuncy, }{weekly or monthly}
}
\value{
df, with either weekly or monthly % growth of cases in each city
}
\description{
view the weekly or monthly number of covid19 cases in Saudi cities
}
