% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cities_first}
\alias{cities_first}
\title{data of first record case in Saudi cities}
\usage{
cities_first()
}
\value{
data frame with first reported cases in each city
}
\description{
This function returns cities with their first reported cases
}
