% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{country_cases}
\alias{country_cases}
\title{KSA covid19 cases}
\usage{
country_cases()
}
\value{
data frame with daily reported cases
}
\description{
use this to view the number of Covid19 cases reported in Saudi arabia
}
