% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{country_cases_growth}
\alias{country_cases_growth}
\title{Growth of Covid 19 cases in Saudi Arabia}
\usage{
country_cases_growth(freqeuncy = "weekly")
}
\arguments{
\item{freqeuncy, }{weekly or monthly}
}
\value{
data frame with either weekly or monthly % growth of cases
}
\description{
view the weekly or monthly number of covid19 cases in Saudi Arabia
}
