% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{country_covid19}
\alias{country_covid19}
\title{KSA covid19 cases}
\usage{
country_covid19()
}
\value{
df, with daily reported cases
}
\description{
use this to view the number of covid19 cases reported in Saudi arabia
}
