% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{country_covid19_growth}
\alias{country_covid19_growth}
\title{Growth of covid 19 cases in Saudi Arabia}
\usage{
country_covid19_growth(freqeuncy = "weekly")
}
\arguments{
\item{freqeuncy, }{weekly or monthly}
}
\value{
df, with either weekly or monthly % growth of cases
}
\description{
view the weekly or monthly number of covid19 cases in Saudi Arabia
}
