% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ksa_shapefiles}
\alias{ksa_shapefiles}
\title{#' shapefiles of saudi cities}
\format{A data frame with 74 cities:
\describe{
\item{city_name}{arabic name of the city}
\item{population}{the city population}
\item{name }{city english name}
\item{geometry }{city polygon}
}}
\source{
\url{https://www.saidgis.com}
}
\usage{
ksa_shapefiles
}
\description{
A dataset containing shapefiles of saudi arabia.
}
\keyword{datasets}
